Token/src/requirements.txt
Flask
pyjwt==1.5.3
gunicorn
